Refactor/Separated Component/Done
